move splitSymbolPairs method to helper
