+++ workers/loc.api/sync/helpers/index.js
@@ -8,6 +8,7 @@ const FOREX_SYMBS = require('./forex-symbs')
 const isForexSymb = require('./is-forex-symb')
 const redirectRequestsToApi = require('./redirect-requests-to-api')
 const checkCollPermission = require('./check-coll-permission')
+const splitSymbolPairs = require('./split-symbol-pairs')
 const {
   delay
 } = require('./utils')
@@ -21,5 +22,6 @@ module.exports = {
   isForexSymb,
   redirectRequestsToApi,
   checkCollPermission,
+  splitSymbolPairs,
   delay
 }
